requirements.txt
pyyaml>=6.0
colorama>=0.4.6
requests>=2.31.0